# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: IA Games Ultimate - Data Labeling Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      parameters:
        - name: level
          in: query
          schema:
            type: string
            enum: [base, pro, expert]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create or register a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/stats:
    get:
      operationId: getUserStats
      tags: [users]
      summary: Get user stats (accuracy, tasks done, streaks, energy)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserStats"
  /users/by-telegram/{telegramId}:
    get:
      operationId: getUserByTelegram
      tags: [users]
      summary: Get user by Telegram ID
      parameters:
        - name: telegramId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found

  # ── Tasks ────────────────────────────────────────────
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [image, text, classification]
        - name: difficulty
          in: query
          schema:
            type: string
            enum: [easy, medium, hard]
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new task (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
  /tasks/next:
    get:
      operationId: getNextTask
      tags: [tasks]
      summary: Get the next available task for a user
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Next task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: No tasks available
...
  /rewards/convert:
    post:
      operationId: convertPoints
      tags: [rewards]
      summary: Convert points to TON (request payout)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConvertPointsBody"
      responses:
        "200":
          description: Conversion result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConvertPointsResult"
  /rewards/missions/{userId}:
    get:
      operationId: getDailyMissions
      tags: [rewards]
      summary: Get daily missions for a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Daily missions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyMission"
  /rewards/energy/recharge:
    post:
      operationId: rechargeEnergy
      tags: [rewards]
      summary: Recharge user energy (via ad or time)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RechargeEnergyBody"
      responses:
        "200":
          description: Energy recharged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserStats"

  # ── Leaderboard ───────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        telegramId:
          type: string
        username:
          type: string
        score:
          type: number
          description: Accuracy score 0-100
        level:
          type: string
          enum: [base, pro, expert]
        points:
          type: integer
        energy:
          type: integer
        maxEnergy:
          type: integer
        xp:
          type: integer
        streak:
          type: integer
        walletAddress:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, telegramId, username, score, level, points, energy, maxEnergy, xp, streak, createdAt]
    CreateUserBody:
      type: object
      properties:
        telegramId:
          type: string
        username:
          type: string
        walletAddress:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
      required: [telegramId, username]
    UpdateUserBody:
      type: object
      properties:
        username:
          type: string
        walletAddress:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
    UserStats:
      type: object
      properties:
        userId:
          type: integer
        tasksCompleted:
          type: integer
        tasksCorrect:
          type: integer
        accuracyRate:
          type: number
        currentStreak:
          type: integer
        energy:
          type: integer
        maxEnergy:
          type: integer
        xp:
          type: integer
        level:
          type: string
          enum: [base, pro, expert]
        points:
          type: integer
        adsWatchedToday:
          type: integer
        dailyTasksDone:
          type: integer
      required: [userId, tasksCompleted, tasksCorrect, accuracyRate, currentStreak, energy, maxEnergy, xp, level, points, adsWatchedToday, dailyTasksDone]

    # ── Task ──────────────────────────────────────────
    Task:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [image, text, classification]
        dataPayload:
          type: object
          description: Task content (image URL, text, or options)
        correctAnswer:
          type: ["string", "null"]
        difficulty:
          type: string
          enum: [easy, medium, hard]
        pointsReward:
          type: integer
        isGolden:
          type: boolean
          description: Golden dataset task with known answer
        consensusCount:
          type: integer
          description: How many responses collected so far
        finalLabel:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, type, dataPayload, difficulty, pointsReward, isGolden, consensusCount, createdAt]
    CreateTaskBody:
      type: object
      properties:
        type:
          type: string
          enum: [image, text, classification]
        dataPayload:
          type: object
        correctAnswer:
          type: ["string", "null"]
        difficulty:
          type: string
          enum: [easy, medium, hard]
        pointsReward:
          type: integer
          default: 10
        isGolden:
          type: boolean
          default: false
      required: [type, dataPayload, difficulty]
...
    ConvertPointsBody:
      type: object
      properties:
        userId:
          type: integer
        points:
          type: integer
      required: [userId, points]
    ConvertPointsResult:
      type: object
      properties:
        success:
          type: boolean
        pointsConverted:
          type: integer
        tonAmount:
          type: number
        remainingPoints:
          type: integer
        message:
          type: string
      required: [success, pointsConverted, tonAmount, remainingPoints, message]
    DailyMission:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [tasks, accuracy, streak, ads]
        target:
          type: integer
        current:
          type: integer
        reward:
          type: integer
        completed:
          type: boolean
        expiresAt:
          type: string
          format: date-time
      required: [id, title, description, type, target, current, reward, completed, expiresAt]
    RechargeEnergyBody:
      type: object
      properties:
        userId:
          type: integer
        method:
          type: string
          enum: [ad, time]
      required: [userId, method]

    # ── Leaderboard ───────────────────────────────────
...