# lib/api-zod/src/generated/types/convertPointsBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IA Games Ultimate - Data Labeling Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ConvertPointsBody {
  userId: number;
  points: number;
}