# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IA Games Ultimate - Data Labeling Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserLevel } from "./userLevel";

export interface User {
  id: number;
  telegramId: string;
  username: string;
  /** Accuracy score 0-100 */
  score: number;
  level: UserLevel;
  points: number;
  energy: number;
  maxEnergy: number;
  xp: number;
  streak: number;
  /** @nullable */
  walletAddress?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: Date;
}